Refactored logic to compute start and end index of formatting in leaf node text.
